API/requirements.txt
streamlit
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters
langchain-google-genai
pypdf
python-dotenv
sentence-transformers
openai
langchain-openai
